+++ test/libsolidity/syntaxTests/inheritance/duplicated_super_constructor_call.sol
@@ -1,0 +1,4 @@
+contract A { constructor(uint) public { } }
+contract B is A(2) { constructor() A(3) public {  } }
+// ----
+// DeclarationError: Duplicated super constructor call.