+++ test/libsolidity/syntaxTests/inheritance/duplicated_super_constructor_call_empty.sol
@@ -1,0 +1,4 @@
+contract A { constructor() public { } }
+contract B is A { constructor() A() public {  } }
+// ----
+// DeclarationError: Duplicated super constructor call.